+++ app/src/main/res/values/strings.xml
@@ -12,5 +12,7 @@
     <string name="menu_settings_title">Настройки</string>
     <string name="page_title">Search it!</string>
     <string name="night_mode">Ночной режим</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 
 </resources>
